+++ android/app/src/main/java/scott/wemessage/app/messages/objects/MessageBase.java
@@ -2,5 +2,5 @@ package scott.wemessage.app.messages.objects;
 
 public abstract class MessageBase {
 
-    public abstract Integer getTimeIdentifier();
+    public abstract Long getTimeIdentifier();
 }
